requirements.txt
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.5
Flask==3.0.2
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1